Split out "physical" part of pool
This part is the part that has the physical storage that the pool then
allocates.

* Move cleanup action that is only physical into physical module
This splits up two very similar functions, but that problem should be
addressed differently than by grouping them in one module.
* Move setup module into physical module. This is the most awkward, as all
functions have to be exposed by the module. But these functions are
intended to be moved out of their current module anyway: see #486.
