+++ src/engine/strat_engine/cleanup.rs
@@ -7,29 +7,8 @@
 use super::super::engine::HasUuid;
 use super::super::errors::{EngineResult, EngineError, ErrorEnum};
 
-use super::blockdev::StratBlockDev;
 use super::pool::StratPool;
 
-/// Wipe some blockdevs of their identifying headers.
-/// Return an error if any of the blockdevs could not be wiped.
-/// If an error occurs while wiping a blockdev, attempt to wipe all remaining.
-pub fn wipe_blockdevs(blockdevs: &[StratBlockDev]) -> EngineResult<()> {
-    let mut unerased_devnodes = Vec::new();
-
-    for bd in blockdevs {
-        let bd_devnode = bd.devnode.to_owned();
-        bd.wipe_metadata()
-            .unwrap_or_else(|_| unerased_devnodes.push(bd_devnode));
-    }
-
-    if unerased_devnodes.is_empty() {
-        Ok(())
-    } else {
-        let err_msg = format!("Failed to wipe already initialized devnodes: {:?}",
-                              unerased_devnodes);
-        Err(EngineError::Engine(ErrorEnum::Error, err_msg))
-    }
-}
 
 /// Teardown pools.
 pub fn teardown_pools(pools: Vec<StratPool>) -> EngineResult<()> {
