+++ src/engine/strat_engine/mod.rs
@@ -2,18 +2,12 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-mod blockdev;
-mod blockdevmgr;
 mod cleanup;
-mod device;
 mod engine;
-mod metadata;
+mod physical;
 mod pool;
 mod serde_structs;
-mod setup;
-mod range_alloc;
 mod thinpool;
-mod util;
 
 pub use self::engine::StratEngine;
