+++ src/engine/strat_engine/physical/device.rs
@@ -13,7 +13,7 @@ use std::path::Path;
 
 use devicemapper::{Bytes, Device, IEC, SECTOR_SIZE, Sectors, devnode_to_devno};
 
-use super::super::errors::{EngineResult, EngineError, ErrorEnum};
+use super::super::super::errors::{EngineResult, EngineError, ErrorEnum};
 
 ioctl!(read blkgetsize64 with 0x12, 114; u64);
