+++ src/engine/strat_engine/pool.rs
@@ -17,10 +17,8 @@ use super::super::engine::{Filesystem, BlockDev, HasName, HasUuid, Pool};
 use super::super::errors::{EngineError, EngineResult, ErrorEnum};
 use super::super::types::{DevUuid, FilesystemUuid, PoolUuid, RenameAction, Redundancy};
 
-use super::blockdevmgr::BlockDevMgr;
-use super::metadata::MIN_MDA_SECTORS;
+use super::physical::{BlockDevMgr, MIN_MDA_SECTORS, get_blockdevs, get_metadata};
 use super::serde_structs::{PoolSave, Recordable};
-use super::setup::{get_blockdevs, get_metadata};
 use super::thinpool::{ThinPool, ThinPoolSizeParams};
 
 pub use super::thinpool::{DATA_BLOCK_SIZE, DATA_LOWATER, INITIAL_DATA_SIZE};
@@ -285,7 +283,7 @@ impl Recordable<PoolSave> for StratPool {
 mod tests {
     use super::super::super::types::Redundancy;
 
-    use super::super::setup::find_all;
+    use super::super::physical::find_all;
     use super::super::tests::{loopbacked, real};
 
     use super::*;
