+++ src/engine/strat_engine/thinpool/thinpool.rs
@@ -20,8 +20,7 @@ use super::super::super::errors::{EngineError, EngineResult, ErrorEnum};
 use super::super::super::structures::Table;
 use super::super::super::types::{DevUuid, PoolUuid, FilesystemUuid, RenameAction};
 
-use super::super::blockdevmgr::{BlockDevMgr, BlkDevSegment, map_to_dm};
-use super::super::device::wipe_sectors;
+use super::super::physical::{BlockDevMgr, BlkDevSegment, map_to_dm, wipe_sectors};
 use super::super::serde_structs::{FilesystemSave, FlexDevsSave, Recordable, ThinPoolDevSave};
 
 use super::dmdevice::{FlexRole, ThinDevIdPool, ThinPoolRole, ThinRole, format_flex_name,
@@ -705,7 +704,7 @@ mod tests {
 
     use devicemapper::{Bytes, SECTOR_SIZE};
 
-    use super::super::super::metadata::MIN_MDA_SECTORS;
+    use super::super::super::physical::MIN_MDA_SECTORS;
     use super::super::super::tests::{loopbacked, real};
     use super::super::super::tests::tempdir::TempDir;
